+++ packages/components/jquery-full/src/index.ts
@@ -1,0 +1,4 @@
+import '@file-viewer/preset-all'
+
+export * from '@file-viewer/jquery'
+export { default } from '@file-viewer/jquery'