+++ packages/components/react-full/src/index.tsx
@@ -1,0 +1,4 @@
+import '@file-viewer/preset-all'
+
+export * from '@file-viewer/react'
+export { default } from '@file-viewer/react'