+++ packages/components/svelte-full/src/index.ts
@@ -1,0 +1,4 @@
+import '@file-viewer/preset-all'
+
+export * from '@file-viewer/svelte'
+export { default } from '@file-viewer/svelte'