+++ packages/components/vue2.7-full/src/index.ts
@@ -1,0 +1,4 @@
+import '@file-viewer/preset-all'
+
+export * from '@file-viewer/vue2.7'
+export { default } from '@file-viewer/vue2.7'